Updating Ceci to these rules:
Components by default broadcast on the "blue" channel. 
Components can be set to broadcast on any single colour
Each component's input can select a color, the default is blue
There are no types
If an input in a component receives a message it can't handle, it logs it to the event console -- this is up to the component developer for now.

<app-counter id="counter">
      <listen on="blue" for="increase">
    </app-counter>

    <app-input id="input">
      <broadcast on="blue">
    </app-input>
      
    <app-label id="label">
      <listen on="blue" for="setValue">
    </app-label>